Update to the latest changes (#4)
* Switch to SNAPSHOT content packages

* Bump webservicesrest to 2.50.0-SNAPSHOT

* (chore) Upgrade Node.js to 22 in frontend Dockerfile (#924)

* Implement the same caching scheme adopted for module-spa

See https://github.com/openmrs/openmrs-module-spa/pull/63 and https://openmrs.atlassian.net/browse/O3-4395. In essence, almost everything is served with Cache-Control set
to no-cache, must-revalidate. This means that any static content that is
not either the openmrs.css file or a Javascript file will make a HTTP
request with If-Modified-Since and If-None-Match headers. The server should
respond to these with a 304 unless they've been changed, which is a fairly
minimal response. Javascript files and openmrs.css are cached for one year.

The idea here is to balance aggressive caching with the need for
implementations to serve custom content from the same webroot.

For implementations not using this image, module-spa implements very similar
caching logic.

* Update to core 2.7.5-SNAPSHOT with a fix for the conceptreferenceranges endpoint

* Update backend modules to newly updated versions

* Bump bedmanagement to 6.2.0-SNAPSHOT

* (test)03-4887: Add support the dispensing app in the reff app e2e distro

* add dispensing e2e

* add dispensing to readme

* README tweaks

* Bump authentication module to 2.0.0-SNAPSHOT (#930)

* Add CACHE_BUST work-around for backend

* Bump FHIR2 to 2.6.0-SNAPSHOT

* Give the backend a little longer to respond

* (fix) Update frontend service queue config

* (release) Release v3.5.0

* (release-revert) Reset to dev versions

---------

Co-authored-by: Ian <52504170+ibacher@users.noreply.github.com>
Co-authored-by: dkayiwa <dkayiwa@openmrs.org>
Co-authored-by: Dennis Kigen <kigen.work@gmail.com>
Co-authored-by: Ian <ian.c.bacher@gmail.com>
Co-authored-by: gitcliff <gitacliff48@gmail.com>
Co-authored-by: chibongho <cbho@pih.org>
Co-authored-by: Jayasanka Weerasinghe <33048395+jayasanka-sack@users.noreply.github.com>
Co-authored-by: openmrs-bot <infrastructure@openmrs.org>